Extract drag handle to own file
